Added syntax highlighting within descriptions
I had some trouble at first including the runmode addon and after a
whole lot of debugging I realized that even though there is a runmode.js
file included at the top of primitive.html, the actual codemirror
instance that was around by the end of the day was the one included in
with the netlogo web instance that got loaded in. So i include the
runmode script at the bottom of the file instead of the top and it
worked beautifully. Really all the script files at the top of
primitive.html could go and I wasn't able to find any difference with
the page, but I'll consult with Umit before taking those out.