Fix jQuery conflict causing search bug
Essentially, using jQuery.noConflict(), you can "save" the instance of
jquery that you currently have loaded and keep it in a variable for
later. So now, even when netlogo loads in its own version of jQuery, we
can still use the one we loaded in by using $local() instead of $().
This way, we can load in typeahead and not have it ever overwritten.